Refactor DNASim CMake setup
This adds a refactored and significantly revised CMake file for the
DNASim library (in DNASim/ directory). This CMake file only deals with
the library sources at the moment (i.e. no tests).

Remarks about the changes:

- Each sub-directory in the DNASim/src directory (e.g. dna/) owns a
CMakeLists.txt file which defines the list of sources in that
sub-directory.
- The DNASim/src CMakeLists.txt file loops over the list of
sub-directories and collect the source files that are part of the
library.
- Within DNASim, headers are now referenced relative to the root of the
source directory (DNASim/src). As a consequence a large part of this
commit is made of changes in the include directives to account for this
new pattern.
- The DNASim/src CMakeLists.txt also includes install directives but
this might need to be adapted.
- A library header (DNASim.h) is now part of the library to simplify the
include process.

In addition, a README is added to the DNASim directory (placeholder as
of now) and install notes are available at the root of the repository in
INSTALL.md

Relates to #3