Remove collision detection code and ODE dependency
Collision detection is not used in emDNA so it is removed from the
codebase. This also makes it possible of getting rid of the ODE (Open
Dynamic Engine) dependency.

This commit also fixes a missing override keyword in `SphereShape`
implementation which was causing a lot of compiler warnings.

Relates to #1