Relocate tclap and force field package
- TCLAP library is relocated to the deps/ directory. There is not GitHub
repo available so a submodule will be difficult to find. - The code
from DNASim/src/dna_force_field_packager is relocated to the root of
the repository. Discussion about how to handle that tool is in #5.

Given that TCLAP was the last dependency to refactor for DNASim this
commit closes #1.