Re-organize emDNA source tree
This commit splits the emDNA sources into the following:

- the emDNA static library sources are in emDNA/src - emDNA command line
tools sources are in emDNA/cli with a directory per cli tool - emDNA
unit tests are in emDNA/tests

In addition, the alglib sources have been relocated to emDNA/deps and a
CMake script has been added to create a alglib static library which is
now included as part of the dependencies.

The TCLAP code base is also re-organized (all sources are now located in
deps/tclap/tclap) to simplify the CMake interface library process.

Finally, the CMake setup for the emDNA static library has been updated
to be on par with that of the DNASim library.

Relates to #3