Incorporate emDNA unit tests to CMake
The emDNA unit tests were not part of the original CMake setup which is
fixed by this commit.

Most unit tests have only been edited to fix include directives
(replaced with a single include for the emDNA static library header).

The BpCollectionHessian test is also fixed: the pre-computed results
were obtained for the Olson1998 force field which was modified in
DNASim; a legacy version of that force field has been added in DNASim
so that the unit test can run successfully.

All unit tests compile and pass.

Relates to #3