Fix issue in SparseMatrixN and update Eigen to 3.3.9
The issue preventing to update to Eigen v3.3.X was a mixing of arrays
and matrices in SparseMatrixN dense/sparse multiplication methods. This
has been fixed by forcing conversion to array when calling the
`set_block` method.

This in turn makes it possible to update the Eigen submodule. The
submodule itself has also been updated to point to the Gitlab repo
rather than the Github one (which does not seem up-to-date). The
project is now tracking Eigen v3.3.9.

All DNASim and emDNA tests pass after the modification.

Closes #6